WMV-363, WMV-364, WMV-366, & WMV-371: Branched off the Core Update to bring VS2017 and 64bit to WMV sooner. -Stole only the best (non-Qt) code from the current Core Update branch! +Known Issue: Loading a model causes WMV to crash.

+++ src/games/wow/HardDriveFile.cpp
@@ -44,7 +44,7 @@ bool HardDriveFile::isAlreadyOpened()
     return false;
 }
 
-bool HardDriveFile::getFileSize(unsigned int & s)
+bool HardDriveFile::getFileSize(size_t & s)
 {
   if (!file && !file->isOpen())
     return false;
